12_OpenAI_Agent_SDK/05_OpenRouter/pyproject.toml
[project]
name = "05-openrouter"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai-agents>=0.0.16",
    "python-dotenv>=1.1.0",
]
